# chan: reduce the memory allocation